Mask move tool, wezel icon, dialog error, import gif, file->new, ROI curve, plot display, user input enhanced
