+++ src/wezel/menu/measure.py
@@ -4,14 +4,15 @@ A collection of methods for measuring numbers from images.
 
 
 from wezel.menu.skimage import VolumeFeatures
-from wezel.menu.scipy import ROIstatistics
+from wezel.menu.scipy import ROIstatistics, ROIcurve
 
 
 
 def all(parent):   
 
+    parent.action(VolumeFeatures, text="3D volume features")
     parent.action(ROIstatistics, text="ROI statistics")
-    parent.action(VolumeFeatures, text="3D Volume Features")
+    parent.action(ROIcurve, text="ROI curve")
